requirements.txt
fastapi~=0.110.1
pydantic~=2.6.4
requests~=2.31.0
uvicorn
gunicorn
psycopg2-binary